% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areaBetweenCurves.R
\name{areaBetweenCurves}
\alias{areaBetweenCurves}
\title{Calculate the area above the bootstrapped thresholds but below the HR 
distribution and vice versa for each point of separation}
\usage{
areaBetweenCurves(survivALL_dfr)
}
\arguments{
\item{survivALL_dfr}{Output of survivALL()}
}
\value{
A per-separation point vector equal to the distance beyond or within
the bootstrapped thresholds
}
\description{
Calculate the area above the bootstrapped thresholds but below the HR 
distribution and vice versa for each point of separation
}
