% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrSignificance.R
\name{hrSignificance}
\alias{hrSignificance}
\title{Calculate HR significance using bootstrap results}
\usage{
hrSignificance(hratios, bs_dfr)
}
\arguments{
\item{hratios}{Hazard ratio vector, output of allHR()}

\item{bs_dfr}{Dataframe of bootstrapped hazard ratios}
}
\value{
A per-separation point vector equal to the distance beyond or within
the bootstrapped thresholds
}
\description{
Calculate HR significance using bootstrap results
}
