% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nki_subset}
\alias{nki_subset}
\title{NKI breast cancer patients subset with complete t.dmfs and e.dmfs information}
\format{An ExpressionSet including assayData with 319 columns (samples), 500
rows (features) and phenoData with 5 columns (variables)}
\source{
http://bioconductor.org/packages/release/data/experiment/html/
breastCancerNKI.html
}
\usage{
nki_subset
}
\description{
An ExpressionSet of NKI breast cancer patients detailing microarray assay and
phenotypic (including survival) information. The data is restricted to the 
500 most variable genes only.
}
\details{
A subset of the breastCancerNKI Bioconductor package - for more details see:

http://bioconductor.org/packages/release/data/experiment/html/
breastCancerNKI.html
}
\keyword{datasets}
